+++ CHANGES.md
@@ -6,6 +6,8 @@ Changes
 
 ### Unreleased
 
+* 2024-03-25 - Backport: Boost Union settings were moved to an admin settings category of its own to support the new theme chooser on Moodle 4.4, resolves #482.
+               This change was backported to this Moodle major version.
 * 2024-03-22 - Backport: Align the before_standard_html_head() function with the new hook callback on Moodle 4.4, resolves #604.
 
 ### v4.1-r23